+++ bin/image-guard.js
@@ -1,10 +1,15 @@
 #!/usr/bin/env node
 
+import { styleText } from 'node:util'
 import { runImageGuard } from '../src/index.js'
 
 try {
   await runImageGuard()
 } catch (err) {
-  console.error('Error running Image Guard:', err.message)
+  if (err.setupFailed) {
+    console.error(styleText('red', err.message))
+  } else {
+    console.error('Error running Image Guard:', err)
+  }
   process.exit(1)
 }